rename scubscription config files
