+++ client/src/app/site/pages/organization/pages/committees/pages/committee-main/components/committee-main/committee-main.component.ts
@@ -4,7 +4,7 @@ import {
     ModelRequestConfig
 } from 'src/app/site/base/base-model-request-handler.component/base-model-request-handler.component';
 
-import { getOrganizationTagListSubscriptionConfig } from '../../../../../organization-tags/config/model-subscription';
+import { getOrganizationTagListSubscriptionConfig } from '../../../../../organization-tags/organization-tags.subscription';
 import { getCommitteeListSubscriptionConfig } from '../../../../committees.subscription';
 
 @Component({
