+++ client/src/app/site/pages/organization/pages/mediafiles/mediafiles.subscription.ts
@@ -1,8 +1,8 @@
 import { map, Observable } from 'rxjs';
 import { Id } from 'src/app/domain/definitions/key-types';
 
-import { ORGANIZATION_ID } from '../../../services/organization.service';
-import { ViewOrganization } from '../../../view-models/view-organization';
+import { ORGANIZATION_ID } from '../../services/organization.service';
+import { ViewOrganization } from '../../view-models/view-organization';
 
 export const ORGANIZATION_MEDIAFILE_LIST_SUBSCRIPTION = `organization_mediafile_list`;
