+++ client/src/app/site/pages/organization/pages/organization-tags/pages/organization-tag-main/components/organization-tag-main/organization-tag-main.component.ts
@@ -4,7 +4,7 @@ import {
     ModelRequestConfig
 } from 'src/app/site/base/base-model-request-handler.component/base-model-request-handler.component';
 
-import { getOrganizationTagListSubscriptionConfig } from '../../../../config/model-subscription';
+import { getOrganizationTagListSubscriptionConfig } from '../../../../organization-tags.subscription';
 
 @Component({
     selector: `os-organization-tag-main`,
